Move _test_nx_conversion to new test_convert module
